se incluyo el tema oscuro al navbar se creo el componente header que incluye al navbar, se creo una excepcion en eslint, se modifico la estructura del navbar
