+++ src/home/pages/HomeScreen.jsx
@@ -1,13 +1,13 @@
 // eslint-disable-next-line no-unused-vars
 import React from 'react'
 import './HomeScreenLogic';
-import Navbarproyect from '../components/Navbarproyect';
+import Header from '../components/Header';
 // import { Navbarproyect } from '../components/Navbarproyect'
 
 export const HomeScreen = () => {
     return (
         <div>
-            <Navbarproyect />
+            <Header />
         </div>
     )
 }
